Add domain/Services/Auth/ManualService & SocialService
