+++ domain/Services/Auth/SocialService.php
@@ -1,12 +1,11 @@
 <?php
 
-namespace Domain\Services;
+namespace Domain\Services\Auth;
 
 use Auth;
 use Exception;
 use Domain\Entities\{
     UserEntity,
-    UserDetailEntity,
     SocialUserAccountEntity
 };
 use Infrastructure\Interfaces\{
@@ -16,10 +15,10 @@ use Infrastructure\Interfaces\{
 use Laravel\Socialite\Contracts\User as SocialUser;
 
 /**
- * Class AuthService
- * @package Domain\Services
+ * Class SocialService
+ * @package Domain\Services\Auth
  */
-class AuthService
+class SocialService
 {
     private $authRepository;
     private $socialRepository;
@@ -35,74 +34,6 @@ class AuthService
     ) {
         $this->authRepository = $authRepository;
         $this->socialRepository = $socialRepository;
-    }
-
-    /**
-     * @param array $oldRequest
-     * @return UserEntity
-     */
-    public function registerUser(array $oldRequest): UserEntity
-    {
-        $userEntity = $this->authRepository->findUser($oldRequest['email']);
-        if (is_null($userEntity)) {
-            $this->authRepository->registerUser($oldRequest);
-            $userEntity = $this->authRepository->findUser($oldRequest['email']);
-        }
-
-        return $userEntity;
-    }
-
-    /**
-     * @param array $oldRequest
-     * @return UserDetailEntity
-     * @throws Exception
-     */
-    public function getUserDetail(array $oldRequest): UserDetailEntity
-    {
-        $userId = $this->authRepository->getUserId($oldRequest);
-        if (is_null($userId)) {
-            throw new Exception('User does not exist');
-        }
-        return $this->authRepository->getUserDetail($userId);
-    }
-
-    /**
-     * @param array $oldRequest
-     * @param int $userId
-     * @return bool
-     */
-    public function login(array $oldRequest, int $userId): bool
-    {
-        $userDetailEntity = $this->authRepository->getUserDetail($userId);
-
-        if ($oldRequest['email'] !== $userDetailEntity->getUserEmail()) {
-            \Log::info("\n【ERROR】Email does not match\n"
-                .'Email:'.$oldRequest['email']
-                .'Password:'.encrypt($oldRequest['password'])
-            );
-            \Log::info($oldRequest['email'].' does not match');
-            return false;
-        }
-
-        if ($oldRequest['password'] !== $userDetailEntity->getPassword()) {
-            \Log::info("\n【ERROR】Password does not match\n"
-                .'Email:'.$oldRequest['email']
-                .'Password:'.encrypt($oldRequest['password'])
-            );
-            return false;
-        }
-
-        // TODO: deleted_at is null ?
-
-        if (!Auth::loginUsingId($userDetailEntity->getUserId(), true)) {
-            \Log::info("\n【ERROR】It is a User that does not exist\n"
-                .'Email:'.$oldRequest['email']
-                .'Password:'.encrypt($oldRequest['password'])
-            );
-            return false;
-        }
-
-        return true;
     }
 
     /**
